doc-ingest-chat/ollama-requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alembic==1.16.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
av==15.0.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chroma==0.2.0
chromadb==1.0.15
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.5
ctranslate2==4.4.0
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
duckdb==1.3.1
durationpy==0.10
einops==0.8.1
fastapi==0.116.0
faster-whisper==1.1.1
filelock==3.18.0
flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
hyperframe==6.1.0
hyperpyyaml==1.2.2
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-core>=0.3.66,<1.0.0
langchain-huggingface==0.3.0
langchain-ollama==0.3.3
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.8
langsmith==0.4.4
lightning==2.5.2
lightning-utilities==0.14.3
llama-cpp-python==0.3.12
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
nltk==3.9.1
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
ollama==0.5.1
omegaconf==2.3.0
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
optuna==4.4.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pdf2image==1.17.0
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pip==25.1.1
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
primepy==1.3
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pyannote-audio==3.3.2
pyannote-core==5.0.0
pyannote-database==5.1.3
pyannote-metrics==3.2.1
pyannote-pipeline==3.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.1
pypika==0.48.9
pyproject-hooks==1.2.0
pytesseract==0.3.13
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.16.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
ruff==0.12.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
semver==3.0.4
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
speechbrain==1.0.3
sqlalchemy==2.0.41
sse-starlette==2.4.1
starlette==0.46.2
starlette-context==0.3.6
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboardx==2.6.4
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.7.4
tqdm==4.67.1
transformers==4.53.1
triton==3.3.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uuid-utils==0.12.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
whisperx==3.4.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
